Completed implementation of L2 error response (which computed the absolute and relative L2 error in the computed solution w.r.t. an analytic exact solution) for Aeras Shallow Water problems. For an example of how to call this response, please see:
examples/Aeras/ShallowWater/TC2/input.xml

The response will actually produce 3 responses (a response of length 3, reported as 3
separate responses in the output).  They are:

1.) The absolute error in the solution w.r.t. the exact solution.
2.) The norm of the exact solution.
3.) The relative error in the solution w.r.t. the exact solution (1.) divided by 2.)).

Right now, only the exact solution for TC2 is coded in the response; other analytical
solutions can be added if desired to the class Aeras::ShallowWaterResponseL2Error.

Note that when the exact solution is 0, 3.) will evaluate to nan or inf b/c of a division by 0
in calculating the relative error.  A warning is printed to the screen in this case.