Modifying mesh for subdivision.
Changed the cubit journal file to include

volume all smooth scheme laplacian free
smooth volume all

the free command allows surface nodes to move during the smoothing process.

The resulting aspect ratio for the worst element was 2.7.

The mesh is subdivied in tet4Subdivision. The subdivision process increases the aspect ratio to 15!

That's not good and a reminder that the quality of the initial tet element discretization
is important.