LCM: THIS COMMIT HAS A NEW FILE: RERUNNING CONFIGURATION SCRIPT MAY BE REQUIRED.
The logic for the creation of pre-inserted surface elements
in the mechanics problem is fairly complex.

Adaptive insertion of surface elements is done by creating an
interface block in the GenericSTKMeshStruct.cpp file. What is there
now is not sufficient to create surface elements: the complex
logic of the mechanics problem is required.

To avoid polluting that file with LCM-specific code, we are going to
extract the surface element initialization logic from the mechanics
problem and move it into LCM Utils.

Then it can be used by both the mechanics problem and GenericSTKMeshStruct.

This commit starts that move.