Bug fix in BasalFrictionCoefficientNode.
This evaluator is needed when one wants to save the computed beta as a node
field (we re-do the same calculations we did at the qp's, but at the nodes).
The evaluator was a doctored copy-and-paste version of the original
BasalFrictionCoefficient, but some lines did not get deleted, and caused
crash due to some field not being created.