Bug fix in DOFSideToCell
Due to (not so recent) changes in Phalanx, we cannot call the method
'dimensions' on a Phalanx field until after setFieldData has been called,
since dimensions checks the underlying Kokkos view (which is not yet set
before setFieldData). Instead, we can call dimensions on the field's
data layout, which is already present and well defined.