Bug fix in StokesFO.
The bug followed the addition of custom equation couplings feature.
In particular, the default equations couplings are built based on the value
of neq at the time the constructor of AbstractProblem is called. If inside
the derived problem one adjusts the number of equations (as it is done in
StokesFO), then one must call setNumEquations, which rebuilds the default
equations couplings (in addition to setting the number of pdes in the rigid
body modes).