A few modifications in some FELIX problems.
Some problems in FELIX (Hydrology, StokesFOHydrology and StokesFO) automatically
register/load/save/gather/scatter states and parameters, based on the
info found in the discretization list. With this commit I added a little more
sugar. In particular, it can happen that the user optimizes over a distributed
parameter but forgets to put it in the discretization list, or puts it, but
lists it only as 'Input' or only as 'Output'. This would cause the parameter
to not be correctly registered (it would not be in the nodalSIS inside the
state manager) and/or to not be gathered/scattered.

To overcome this, we need to inspect the optimization sublist inside the
solver options, to check if (1) we are using ROL (I haven't implemented
the other optimizers yet), and (2) what are the distributed parameters we
are optimizing upon (if any). If we find some, we enforce two things:

1) if we find them in the discretization sublist(s), we mark the field as
   both input and output, so that later on the gather/scatter evaluators
   are indeed built
2) if we do not find them in the discretization sublist(s), we manually
   register them in the state manager, and create their gather/scatter
   evaluators.